% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_combos}
\alias{make_combos}
\title{Make a comprehensive data frame with all combinations of CO2 and I}
\usage{
make_combos(CO2_v, n, p)
}
\arguments{
\item{CO2_v}{a vector of CO2 values}

\item{n}{the number of individuals}

\item{p}{the prevalence in percent of infections in your area}
}
\description{
Make a comprehensive data frame with all combinations of CO2 and I
}
