% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pi_df}
\alias{pi_df}
\title{Create a data frame of pi values at different CO2 concentrations}
\usage{
pi_df(CO2_v, M, q, t, n, nu_out, nu_in, Ip, C0 = 415, p = 0.48)
}
\arguments{
\item{CO2_v}{The vector of possible CO2 concentrations}

\item{M}{Emission rate of CO2 (m^3/h/person)}

\item{q}{Generation rate of infectious quanta (/h/person)}

\item{t}{Length of exposure (h)}

\item{n}{Number of individuals in the room (person)}

\item{nu_out}{Exhalation filtration efficiency (no units)}

\item{nu_in}{Inhalation filtration efficiency (no units)}

\item{Ip}{the prevalence of covid in the community in percent (0 - 100)}

\item{C0}{Outdoor CO2 concentration (ppm)}

\item{p}{pulmonary ventilation rate of a person (m^3/h)}
}
\description{
Create a data frame of pi values at different CO2 concentrations
}
