% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_pi}
\alias{plot_pi}
\title{Plot the risk line graph}
\usage{
plot_pi(df_plt, C, M, q, t, n, nu_out, nu_in, Ip, C0 = 415, p = 0.48)
}
\arguments{
\item{C}{Current CO2 value}

\item{M}{Emission rate of CO2 (m^3/h/person)}

\item{q}{Generation rate of infectious quanta (/h/person)}

\item{t}{Length of exposure (h)}

\item{n}{Number of individuals in the room (person)}

\item{nu_out}{Exhalation filtration efficiency (no units)}

\item{nu_in}{Inhalation filtration efficiency (no units)}

\item{Ip}{the prevalence of covid in the community in percent (0 - 100)}

\item{C0}{Outdoor CO2 concentration (ppm)}

\item{p}{pulmonary ventilation rate of a person (m^3/h)}

\item{df}{A data frame with columns CO2 and p_i}
}
\description{
Plot the risk line graph
}
