% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prob_inf}
\alias{prob_inf}
\title{Calculate the probability of infection at a given I and CO2}
\usage{
prob_inf(C, M, I, q, t, n, nu_out, nu_in, C0, p)
}
\arguments{
\item{C}{The indoor CO2 concentration (ppm)}

\item{M}{Emission rate of CO2 (m^3/h/person)}

\item{I}{Number of infectious individuals in the room (person)}

\item{q}{Generation rate of infectious quanta (/h/person)}

\item{t}{Length of exposure (h)}

\item{n}{Number of individuals in the room (person)}

\item{nu_out}{Exhalation filtration efficiency (no units)}

\item{nu_in}{Inhalation filtration efficiency (no units)}

\item{C0}{Outdoor CO2 concentration (ppm)}

\item{p}{pulmonary ventilation rate of a person (m^3/h)}
}
\description{
Calculate the probability of infection at a given I and CO2
}
